general wip (need to get latest work on server before go home!)

+++ src/SFA.DAS.ProviderRelationships.Api.Client/Http/RestClientException.cs
@@ -40,7 +40,7 @@ namespace SFA.DAS.ProviderRelationships.Api.Client.Http
         private static string GenerateMessage(HttpResponseMessage httpResponseMessage, string errorResponse)
         {
             return
-$@"Request '{httpResponseMessage.RequestMessage.RequestUri}' returned {httpResponseMessage.StatusCode} {httpResponseMessage.ReasonPhrase}
+$@"Request '{httpResponseMessage.RequestMessage.RequestUri}' returned {(int)httpResponseMessage.StatusCode} {httpResponseMessage.ReasonPhrase}
 Response:
 {errorResponse}";
         }
